tests ok

+++ migrations/sqls/20220907215919-orders-up.sql
@@ -1,6 +1,6 @@
 CREATE TABLE orders (
     id serial PRIMARY KEY,
     status varchar(15) default 'open',
-    user_id bigint REFERENCES users (id)
+    user_id INT REFERENCES users(id)
 );
