+++ migrations/sqls/20220907215931-order-products-up.sql
@@ -1,8 +1,8 @@
 CREATE TABLE order_products (
     quantity INT not null default 1,
-    order_id bigint ,
+    order_id INT ,
     foreign key (order_id) references orders(id),
-    product_id bigint ,
+    product_id INT ,
     foreign key (product_id) references products(id),
     primary key (order_id, product_id)
 );
